+++ src/validators/swagger_request_models.py
@@ -46,6 +46,7 @@ from helpers.constants import (
     RuleSourceType,
     ServiceOperationType,
 )
+from helpers import Version, NextToken
 from helpers.regions import AllRegions, AllRegionsWithGlobal
 from helpers.time_helper import utc_datetime
 from models.rule import RuleIndex
@@ -86,6 +87,12 @@ class BasePaginationModel(BaseModel):
         None,
         description='Provide next_token received from the previous request',
     )
+
+    @field_validator('next_token', mode='after')
+    @classmethod
+    def _validate_next_token(cls, next_token):
+        NextToken.deserialize(next_token)
+        return next_token
 
 
 class TimeRangedMixin:
@@ -1150,6 +1157,11 @@ class JobPostModel(BaseModel):
         None,
         description='License to exhaust for this job. Will be resolved '
         'automatically unless an ambiguous occurs',
+    )
+
+    application_id: str = Field(
+        None,
+        description='Application ID with credentials for this job',
     )
 
     dojo_product: str = Field(